restrict activities page to admins
