+++ src/app/forms/[id]/page.tsx
@@ -1,7 +1,18 @@
+import { getServerSession } from 'next-auth';
+import { redirect } from 'next/navigation';
+import { authOptions } from '@/lib/auth';
 import { prisma } from '@/lib/prisma';
 import FormDisplay from './form';
 
 export default async function FormPage({ params }: { params: { id: string } }) {
+  const session = await getServerSession(authOptions);
+  if (
+    !session ||
+    (session.user.role !== 'ADMIN' && session.user.role !== 'SUPER_ADMIN')
+  ) {
+    redirect('/');
+  }
+
   const form = await prisma.form.findUnique({
     where: { id: params.id },
     include: { fields: { orderBy: { order: 'asc' } } },
