--- src/middleware.ts
@@ -1,30 +1,0 @@
-import { NextResponse } from 'next/server';
-import type { NextRequest } from 'next/server';
-import { getToken } from 'next-auth/jwt';
-
-export async function middleware(req: NextRequest) {
-  const token = await getToken({ req, secret: process.env.NEXTAUTH_SECRET });
-  const { pathname } = req.nextUrl;
-
-  if (pathname.startsWith('/admin/site')) {
-    if (!token || token.role !== 'SUPER_ADMIN') {
-      return NextResponse.redirect(new URL('/', req.url));
-    }
-  } else if (pathname.startsWith('/admin') || pathname.startsWith('/forms')) {
-    if (!token || (token.role !== 'ADMIN' && token.role !== 'SUPER_ADMIN')) {
-      return NextResponse.redirect(new URL('/', req.url));
-    }
-  }
-
-  if (pathname.startsWith('/members')) {
-    if (!token) {
-      return NextResponse.redirect(new URL('/api/auth/signin', req.url));
-    }
-  }
-
-  return NextResponse.next();
-}
-
-export const config = {
-  matcher: ['/admin/:path*', '/members/:path*', '/forms/:path*'],
-};
